% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLogFC.R
\name{calcLogFC}
\alias{calcLogFC}
\title{calcLogFC}
\usage{
calcLogFC(expr, annot, var)
}
\arguments{
\item{expr}{expr data from obj slot}

\item{annot}{metadata}

\item{var}{names of variable/cell-type to consider}
}
\description{
calculate LogFC of expr
}
\examples{
expr= obj@assays$RNA@data[varfeat,]
annot= obj@meta.data
var='macrophage'
calcLogFC(expr,annot,var)
}
