% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSim_avgCT.R
\name{calcSim_avgCT}
\alias{calcSim_avgCT}
\title{calcSim_avgCT}
\usage{
calcSim_avgCT(df, annot, id.list, ct, res_dir)
}
\arguments{
\item{df}{dataframe of cell embeddings}

\item{annot}{metadata}

\item{id.list}{ids for the first and second datasets}

\item{ct}{cell-type}

\item{res_dir}{results directory}
}
\description{
calculate average similarity between two datasets based on CCA info for each cell-type
}
\examples{
calcSim_avgCT(df, annot, id.list,df_type, ct, res_dir)
}
