% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGexprDensity.R
\name{checkGexprDensity}
\alias{checkGexprDensity}
\title{checkGexprDensity}
\usage{
checkGexprDensity(obj, dataset_name, ct, res_dir)
}
\arguments{
\item{obj}{object}

\item{dataset_name}{dataset name}

\item{ct}{cell-type}

\item{res_dir}{results directory to save plot}
}
\description{
check Gene expression Density
}
\examples{
expr_dir = '/Users/rao198/OneDrive\ -\ Harvard\ University/Informatics/doc/paper/DA4LT/V1/analysis/brain/'
res_dir =  paste0(expr_dir, '01_processData/')
obj = readRDS(paste0(res_dir,'mouse_sc_commonVarG.RDS'))  #commonVarGmouse_sc.RDS
ct='Chandelier'
dataset_name = 'mouse'
checkGexprDensity(obj,dataset_name,ct,res_dir)
}
