% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDir.R
\name{createDir}
\alias{createDir}
\title{createDir}
\usage{
createDir(dirName)
}
\arguments{
\item{dirName}{directory folder name}
}
\description{
create directory folder if it does not already exists
}
\examples{
dirName='mainDir'
createDir(dirName)
}
