% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2obj.R
\name{df2obj}
\alias{df2obj}
\title{df2obj}
\usage{
df2obj(df, annot)
}
\arguments{
\item{df}{dataframe}

\item{annot}{metadata}
}
\value{
obj object
}
\description{
convert datframe to obj
}
\examples{
df2obj(df,annot)
}
