% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonGenesPerPathway.R
\name{getCommonGenesPerPathway}
\alias{getCommonGenesPerPathway}
\title{getCommonGenesPerPathway}
\usage{
getCommonGenesPerPathway(rankedPathways, enrichedG_1, enrichedG_2)
}
\arguments{
\item{rankedPathways}{dataframe of ranked pathways from 07 step}

\item{enrichedG_1}{genes enriched per pathway for dataset 1}

\item{enrichedG_2}{genes enriched per pathway for dataset 2}
}
\value{
p_g_common common genes per pathway
}
\description{
get gene members common per pathway between two datasets
}
\examples{
getCommonGenesPerPathway(rankedPathways,enrichedG_1,enrichedG_2)
}
