% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonPathways.R
\name{getCommonPathways}
\alias{getCommonPathways}
\title{getCommonPathways}
\usage{
getCommonPathways(p1, p2)
}
\arguments{
\item{p1}{CCA res}

\item{p2}{metadata}
}
\value{
commpnP common set of pathways
}
\description{
get common  pathways between two lists of pathways
}
\examples{
run_07_rankPathways()
}
