% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichedGenes.R
\name{getEnrichedGenes}
\alias{getEnrichedGenes}
\title{getEnrichedGenes}
\usage{
getEnrichedGenes(gsea_res, rankedPathways)
}
\arguments{
\item{gsea_res}{gsea_res res}

\item{rankedPathways}{ranked pathways data.frame}
}
\value{
enrichedG ranked pathways with info
}
\description{
get gene members in pathways enriched in samples of interest based on gsea_res results
}
\examples{
getEnrichedGenes(gsea_res,rankedPathways)
}
