% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadataStat.R
\name{getMetadataStat}
\alias{getMetadataStat}
\title{getMetadataStat}
\usage{
getMetadataStat(annot, opfname)
}
\arguments{
\item{metadata}{metadata}
}
\value{
metadata_updated updated metadata
}
\description{
get Stat of metadata
}
\examples{
getMetadataStat(metadata)
}
