% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOverlapMarkers.R
\name{getOverlapMarkers}
\alias{getOverlapMarkers}
\title{getOverlapMarkers}
\usage{
getOverlapMarkers(markers1, id1, markers2, id2, res_dir)
}
\arguments{
\item{id1}{id of markers1}

\item{id2}{id of markers2}

\item{res_dir}{results_directory to save results and plot}

\item{Markers1}{Markers of obj1}

\item{Markers2}{Markers of obj2}
}
\description{
get Overlap Markers between two objs
}
\examples{
getOverlapMarkers(markers1,markers2,res_dir)
}
