% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathwaysInfo.R
\name{getPathwaysInfo}
\alias{getPathwaysInfo}
\title{getPathwaysInfo}
\usage{
getPathwaysInfo(pathwaysToAnnotate)
}
\arguments{
\item{pathwaysToAnnotate}{set of pathways to annotate}
}
\value{
P_annnot data.frame of pathways info
}
\description{
merge sample scores from two datasets
}
\examples{
getPathwaysInfo(pathwaysToAnnotate)
}
