% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeMultiFiles.R
\name{mergeMultiFiles}
\alias{mergeMultiFiles}
\title{mergeMultiFiles}
\usage{
mergeMultiFiles(my.files)
}
\arguments{
\item{my.files}{sample scores files}
}
\value{
mergedFile merged file of all files
}
\description{
merge sample scores from one dataset
}
\examples{
mergeMultiFiles(my.files)
}
