% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performMarkersAnalysis.R
\name{performMarkersAnalysis}
\alias{performMarkersAnalysis}
\title{performMarkersAnalysis}
\usage{
performMarkersAnalysis(markers, dataset_id, res_dir)
}
\arguments{
\item{dataset_id}{dataset_id of obj}

\item{res_dir}{results_directory to save results and plot}

\item{Markers}{Markers of obj}
}
\description{
perform Analysis using Markers of obj
}
\examples{
performMarkersAnalysis(Markers,opfname)
}
