% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCombineRankingPathways.R
\name{plotCombineRankingPathways}
\alias{plotCombineRankingPathways}
\title{plotCombineRankingPathways}
\usage{
plotCombineRankingPathways(fname.lst, ct.lst, res_dir)
}
\arguments{
\item{fname.lst}{list of ranking pathways files}

\item{ct.lst}{list of cell-types according to the files}

\item{res_dir}{result directory to save the plot}
}
\description{
combine rankingP across all cell-types
}
\examples{
plotCombineRankingPathways(fname.lst,ct.lst,res_dir)
}
