% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRankingG.R
\name{plotRankingG}
\alias{plotRankingG}
\title{plotRankingG}
\usage{
plotRankingG(df, col_annot1, col_annot2, title, opfname)
}
\arguments{
\item{df}{dataframe to plot}

\item{col_annot1}{first column annotation}

\item{col_annot2}{second column annotation}

\item{title}{title of plot}

\item{opfname}{file to print plot}
}
\description{
plot ranking genes per pathway
}
\examples{
plotRankingG(df,col_annot1,col_annot2,title,opfname)
}
