% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRankingP.R
\name{plotRankingP}
\alias{plotRankingP}
\title{plotRankingP}
\usage{
plotRankingP(df, col_annot1, col_annot2, row_annot1, title, opfname)
}
\arguments{
\item{df}{dataframe to plot}

\item{col_annot1}{first column annotation}

\item{col_annot2}{second column annotation}

\item{row_annot1}{first row annotation}

\item{title}{title of plot}

\item{opfname}{file to print plot}
}
\description{
plot pathways ranking
}
\examples{
plotRankingP(df,col_annot1,col_annot2,row_annot1,title,opfname)
}
