% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCCAres_allCT.R
\name{processCCAres_allCT}
\alias{processCCAres_allCT}
\title{processCCAres_allCT}
\usage{
processCCAres_allCT(cca_res, x, obj1, y, obj2)
}
\arguments{
\item{cca_res}{CCA res}

\item{annot}{metadata}
}
\value{
cca_res_updated updated cca_res with info
}
\description{
process CCA res to incluide extra info from all cell-types
}
\examples{
processCCAres_allCT(df,annot)
}
