% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCCAres_perCT.R
\name{processCCAres_perCT}
\alias{processCCAres_perCT}
\title{processCCAres_perCT}
\usage{
processCCAres_perCT(cca_res, x, y)
}
\arguments{
\item{cca_res}{CCA res}

\item{x}{dataframe score of dataset 1}

\item{y}{dataframe score of dataset 2}
}
\value{
cca_res cca results updated for this x & y ct-based datasets
}
\description{
get ct-specific info from CCA res
}
\examples{
processCCAres_perCT(cca_res,x,y)
}
