% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processIntegratedres_allCT.R
\name{processIntegratedres_allCT}
\alias{processIntegratedres_allCT}
\title{processIntegratedres_allCT}
\usage{
processIntegratedres_allCT(integrated_res, x, obj1, y, obj2)
}
\arguments{
\item{integrated_res}{CCA res}

\item{x}{metadata}
}
\value{
integrated_res_updated updated integrated_res with info
}
\description{
process integration res to incluide extra info from all cell-types
}
\examples{
processIntegratedres_allCT(df,annot)
}
