% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processObj.R
\name{processObj}
\alias{processObj}
\title{processObj}
\usage{
processObj(obj)
}
\arguments{
\item{obj}{obj data}
}
\value{
obj processed obj data
}
\description{
process obj using Seurat workflow
}
\examples{
obj=x
processObj(obj)
}
