% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPathwaysCorrelation.R
\name{rankPathwaysCorrelation}
\alias{rankPathwaysCorrelation}
\title{rankPathwaysCorrelation}
\usage{
rankPathwaysCorrelation(cca_res, p.info, w_opt)
}
\arguments{
\item{cca_res}{CCA res}

\item{p.info}{pathways info}

\item{w_opt}{to use weighted pathways scores}

\item{id.list}{id names for the objects}
}
\value{
rankP.cor  ranked pathways based on correlation
}
\description{
rank pathways based on correlation
}
\examples{
rankPathwaysCorrelation(cca_res,p.info,w_opt)
}
