% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnamingCTconv.R
\name{RnamingCTconv}
\alias{RnamingCTconv}
\title{RnamingCTconv}
\usage{
RnamingCTconv(obj)
}
\arguments{
\item{obj}{obj data}
}
\value{
obj updated obj data
}
\description{
Rnaming cell-type labels convienent for R
}
\examples{
obj=x
RnamingCTconv(obj)
}
