% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_00_preprocessData.R
\name{run_00_preprocessData}
\alias{run_00_preprocessData}
\title{run_00_preprocessData}
\usage{
run_00_preprocessData(
  data_dir,
  res_dir,
  id.list,
  obj.fname.list,
  convertG.list,
  processObjOpt,
  typeColumnToUse.list,
  assaySlotToUse.list
)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{id.list}{id names for the objects}

\item{obj.fname.list}{file names for the objects
#' @param convertG.list to convert genes to human T,F}

\item{processObjOpt}{T or F to process object}

\item{typeColumnToUse.list}{list of column name holds the cell-type labels}

\item{assaySlotToUse.list}{list of name of assay slot}
}
\description{
preprocess data
}
\examples{
run_00_preprocessData(data_dir,res_dir, id.list,obj.fname.list, convertG.list,processObjOpt,
typeColumnToUse.list,assaySlotToUse.list)
}
