% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_01_processData.R
\name{run_01_processData}
\alias{run_01_processData}
\title{run_01_processData}
\usage{
run_01_processData(data_dir, res_dir, id.list, obj.fname.list, genes2keep)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{id.list}{id names for the objects}

\item{obj.fname.list}{file names for the objects}

\item{genes2keep}{which genes to keep between the two objects}
}
\value{
processedData list of processedData: obj1,obj2, and annot
}
\description{
process data
}
\examples{
run_01_processData(data_dir,res_dir, id.list=c(id1,id2),obj.fname.list =c(obj1_fname,obj2_fname),genes2keep)
}
