% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_02_gseaFast.R
\name{run_02_gseaFast}
\alias{run_02_gseaFast}
\title{run_02_gseaFast}
\usage{
run_02_gseaFast(
  data_dir,
  res_dir,
  dataset_name,
  expr,
  annot,
  dataType,
  geneType,
  pathwaydatabase,
  ct
)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{dataset_name}{id name for the object}

\item{expr}{data from object}

\item{annot}{annot from object}

\item{dataType}{which obj data slot to use}

\item{geneType}{which genes to use}

\item{pathwaydatabase}{pathway database}

\item{ct}{name of cell-type}
}
\description{
score samples with gsea for a single cell-type at a time
}
\examples{
run_02_gseaFast(data_dir,res_dir, dataset_name,expr,annot,dataType,geneType,pathwaydatabase,ct)
}
