% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_03_sampleScoring.R
\name{run_03_sampleScoring}
\alias{run_03_sampleScoring}
\title{run_03_sampleScoring}
\usage{
run_03_sampleScoring(GSEA_res, filter_opt, filter_cutoff, sample_cutoff)
}
\arguments{
\item{GSEA_res}{gsea results}

\item{filter_opt}{column of GSEA to filter based on i.e., pvalue}

\item{filter_cutoff}{0.05}

\item{sample_cutoff}{percentage of samples meets the cutoff of filter i.e. 0.1}
}
\description{
score and filter samples based on gsea results
}
\examples{
run_03_sampleScoring(GSEA_res,filter_opt,filter_cutoff,sample_cutoff)
}
