% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_05_CCA.R
\name{run_05_CCA}
\alias{run_05_CCA}
\title{run_05_CCA}
\usage{
run_05_CCA(obj1, obj2, num.cc)
}
\arguments{
\item{obj1}{object of dataset 1}

\item{obj2}{object of dataset 2}
}
\value{
ccaRes CCA results
}
\description{
run Seurat CCA over two objectes
}
\examples{
run_05_CCA(obj1,obj2)
}
