% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_05_dataIntegration.R
\name{run_05_dataIntegration}
\alias{run_05_dataIntegration}
\title{run_05_dataIntegration}
\usage{
run_05_dataIntegration(obj1, obj2)
}
\arguments{
\item{obj1}{object of dataset 1}

\item{obj2}{object of dataset 2}
}
\value{
integrated_res  integration obj
}
\description{
run Seurat data integarte over two objectes
}
\examples{
run_05_dataIntegration(obj1,obj2)
}
