% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_07_rankPathways.R
\name{run_07_rankPathways}
\alias{run_07_rankPathways}
\title{run_07_rankPathways}
\usage{
run_07_rankPathways(cca_res, p.info, w_opt)
}
\arguments{
\item{cca_res}{CCA res}

\item{p.info}{pathway info}

\item{w_opt}{to use weighted pathways scores}
}
\value{
rankP ranked pathways with info
}
\description{
rank pathways
}
\examples{
run_07_rankPathways(cca_res,p.info,w_opt)
}
