% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_08_rankGenes.R
\name{run_08_rankGenes}
\alias{run_08_rankGenes}
\title{run_08_rankGenes}
\usage{
run_08_rankGenes(expr1, expr2, commonG, rankedPathways)
}
\arguments{
\item{expr1}{gene expression or logFC of dataset 1}

\item{expr2}{gene expression or logFC of dataset 2}

\item{commonG}{common genes per pathway between the two datasets}

\item{rankedPathways}{ranked pathways with info}
}
\value{
rankG ranked genes with info
}
\description{
rank genes in rankedPathways based on the expression of genes between the two datasets
}
\examples{
run_08_rankGenes(expr1,expr2,commonG,rankedrankedPathways)
}
