% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_09_detectUnknownCT.R
\name{run_09_detectUnknownCT}
\alias{run_09_detectUnknownCT}
\title{run_09_detectUnknownCT}
\usage{
run_09_detectUnknownCT(obj1, obj2, annot, id.list, ct, res_dir)
}
\arguments{
\item{obj1}{obj of the first dataset}

\item{obj2}{obj of the second dataset}

\item{annot}{metadata}

\item{id.list}{name list of the two datasets}

\item{ct}{cell-type to detect}

\item{res_dir}{results directory}
}
\description{
detect unknown cell-type between two datasets (Reference and Query)
}
