% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDA4CA.R
\name{runDA4CA}
\alias{runDA4CA}
\title{runDA4CA}
\usage{
runDA4CA()
}
\arguments{
\item{expr_dir}{experiment directory folder name}

\item{id.list}{id names for the objects}

\item{obj.fname.list}{file names for the objects}

\item{typeColumnToUse.list}{list of column name holds the cell-type labels}

\item{assaySlotToUse.list}{list of name of assay slot}

\item{genes2keep}{which genes to keep between the two objects}

\item{processObjOpt}{T or F to process object with visualization}

\item{pathwaydatabase}{pathway database}

\item{dataType}{which obj data slot to use}

\item{geneType}{which genes to use}

\item{filter_opt}{column of GSEA to filter based on i.e., pvalue}

\item{filter_cutoff}{0.05}

\item{sample_cutoff}{percentage of samples meets the cutoff of filter i.e. 0.1}

\item{topRanked}{10 top ranked biological findings to investigate, used for pathways & genes}
}
\description{
run DA4CA main function to run the analysis using two objects
}
\examples{
data_name = 'panc'
expr_dir =  paste0('/Users/rao198/OneDrive\ -\ Harvard\ University/Informatics/doc/paper/DA4LT/V1/analysis/',data_name)
createDir(expr_dir)
id.list=c('mouse','human')
obj.fname.list=c('mouse_sc.RDS','human_sc.RDS')
typeColumnToUse.list=c('cell.names','cell.names')
assaySlotToUse.list=list('integrated','SCT')
genes2keep = 'commonVarG'
processObjOpt = F
pathwaydatabase = 'reactome.db'
dataType = 'norm'
geneType = 'var'
filter_opt = 'pvalue'
filter_cutoff = 0.05
sample_cutoff = 0.07
topRanked = 10
runDA4CA(expr_dir, id.list,obj.fname.list,typeColumnToUse.list,assaySlotToUse.list,
genes2keep,processObjOpt,
pathwaydatabase,dataType,geneType,
filter_opt,filter_cutoff,sample_cutoff,
topRanked)
}
