% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSEA.R
\name{runGSEA}
\alias{runGSEA}
\title{run_02_gsea}
\usage{
runGSEA(pathwaydatabase, DE_lst, reference_samples)
}
\arguments{
\item{pathwaydatabase}{pathway database}

\item{DE_lst}{list of logFCs}

\item{reference_samples}{control samples}
}
\description{
score samples with gsea
}
\examples{
run_02_gsea(pathwaydatabase,DE_lst,reference_samples)
}
