% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCCA.R
\name{selectCCA}
\alias{selectCCA}
\title{selectCCA}
\usage{
selectCCA(cca_res, select_opt, select_opt_cutoff)
}
\arguments{
\item{select_opt}{option to select dims}

\item{select_opt_cutoff}{cutoff to select dims}

\item{cor}{CCA cor}
}
\value{
selected_dims selected dims to use
}
\description{
select important CCA dims
}
\examples{
selectCCA(cca_res,select_opt,select_opt_cutoff)
}
