% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubsetObjVarF.R
\name{SubsetObjVarF}
\alias{SubsetObjVarF}
\title{df2objVarF}
\usage{
SubsetObjVarF(obj, VarF, cells)
}
\arguments{
\item{obj}{object}

\item{VarF}{list variable features}

\item{cells}{list cells}
}
\value{
obj object
}
\description{
convert datframe to obj with selected variable features
}
\examples{
df2objVarF(df,annot,VarF)
}
