% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visData.R
\name{visData}
\alias{visData}
\title{visData}
\usage{
visData(df, annot, p.df, opfname)
}
\arguments{
\item{df}{dataframe}

\item{annot}{metadata for df}

\item{p.df}{dataframe holding pathways info}

\item{opfname}{file name to save plot}
}
\description{
visualize dataframe
}
\examples{
obj=x
processObj(obj)
}
