% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visObj.R
\name{visObj}
\alias{visObj}
\title{visObj}
\usage{
visObj(obj, opfname)
}
\arguments{
\item{obj}{object}

\item{opfname}{file name to save plot}
}
\description{
visualize obj using Seurat
}
\examples{
visObj(obj,opfname)
}
