% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommonPathways.R
\name{getCommonPathways}
\alias{getCommonPathways}
\title{getCommonPathways}
\usage{
getCommonPathways(p1, p2)
}
\arguments{
\item{p1}{list of pathways from dataset 1 selected for cell-type x}

\item{p2}{list of pathways from dataset 2 selected for cell-type y}
}
\value{
commpnP common set of pathways
}
\description{
get common  pathways between two lists of pathways
}
\examples{
getCommonPathways(p1,p2)
}
