% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariableFeatures.R
\name{getVariableFeatures}
\alias{getVariableFeatures}
\title{getVariableFeatures}
\usage{
getVariableFeatures(sc, no_feat)
}
\arguments{
\item{no_feat}{number of features}

\item{sc_obj}{obj}
}
\value{
name of method used to select fetaures
}
\description{
get method of variable features from obj
}
\examples{
getVariableFeatures(sc_obj,no_feat)
}
