% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetadataStat.R
\name{plotMetadataStat}
\alias{plotMetadataStat}
\title{plotMetadataStat}
\usage{
plotMetadataStat(annot, res_dir)
}
\arguments{
\item{annot}{metadata}

\item{res_dir}{results directory to save plot}
}
\description{
plot Stat of metadata
}
\examples{
plotMetadataStat(annot,opfname)
}
