% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_00_preprocessData.R
\name{run_00_preprocessData}
\alias{run_00_preprocessData}
\title{run_00_preprocessData}
\usage{
run_00_preprocessData(
  data_dir,
  res_dir,
  id.list,
  obj.fname.list,
  convertG.list,
  processObjOpt,
  typeColumnToUse.list,
  assaySlotToUse.list
)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{id.list}{id names for the objects}

\item{obj.fname.list}{file names for the objects}

\item{convertG.list}{to convert genes to human T,F}

\item{processObjOpt}{T or F to process object}

\item{typeColumnToUse.list}{list of column name holds the cell-type labels}

\item{assaySlotToUse.list}{list of name of assay slot}
}
\description{
preprocess data
}
\examples{
data_name = 'panc'
expr_dir =  paste0('/Users/rao198/OneDrive\ -\ Harvard\ University/Informatics/doc/paper/DA4LT/V1/analysis/',data_name)
data_dir =  paste0(expr_dir,'/data/')
res_dir =  paste0(expr_dir,'/01_processData/')
createDir(res_dir)
id.list=c('mouse','human')
obj.fname.list=c('mouse_sc.RDS','human_sc.RDS')
typeColumnToUse.list=c('cell.names','cell.names') # later replaced by 'type'
assaySlotToUse.list=list('integrated','SCT')
run_00_preprocessData(data_dir,res_dir, id.list,obj.fname.list,typeColumnToUse.list,assaySlotToUse.list)
}
