% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_01_processData.R
\name{run_01_processData}
\alias{run_01_processData}
\title{run_01_processData}
\usage{
run_01_processData(data_dir, res_dir, id.list, obj.fname.list, genes2keep)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{id.list}{id names for the objects}

\item{obj.fname.list}{file names for the objects}

\item{genes2keep}{which genes to keep between the two objects}
}
\value{
processedData list of processedData: obj1,obj2, and annot
}
\description{
process data
}
\examples{
data_name = 'panc'
expr_dir =  paste0('/Users/rao198/OneDrive\ -\ Harvard\ University/Informatics/doc/paper/DA4LT/V1/analysis/',data_name)
data_dir =  paste0(expr_dir,'/data/')
res_dir =  paste0(expr_dir,'/01_processData/')
createDir(res_dir)
id1='mouse'
id2='human'
obj1_fname='mouse_sc.RDS'
obj2_fname='human_sc.RDS'
genes2keep='commonVarG'
run_01_processData(data_dir,res_dir, id.list=c(id1,id2),obj.fname.list =c(obj1_fname,obj2_fname),genes2keep)
}
