% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_02_gsea.R
\name{run_02_gsea}
\alias{run_02_gsea}
\title{run_02_gsea}
\usage{
run_02_gsea(
  data_dir,
  res_dir,
  dataset_name,
  expr,
  annot,
  dataType,
  geneType,
  pathwaydatabase
)
}
\arguments{
\item{data_dir}{data directory folder name}

\item{res_dir}{result directory folder name}

\item{dataset_name}{id name for the object}

\item{expr}{data from object}

\item{annot}{annot from object}

\item{dataType}{which obj data slot to use}

\item{geneType}{which genes to use}

\item{pathwaydatabase}{pathway database}
}
\description{
score samples with gsea for serial run on cell-types of an obj
}
\examples{
run_02_gsea(data_dir,res_dir, dataset_name,expr,annot,dataType,geneType,pathwaydatabase)
}
