% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_04_mergeTwoDatasets.R
\name{run_04_mergeTwoDatasets}
\alias{run_04_mergeTwoDatasets}
\title{run_04_mergeTwoDatasets}
\usage{
run_04_mergeTwoDatasets(df1, df2)
}
\arguments{
\item{df1}{merged scoring files for dataset 1}

\item{df2}{merged scoring files for dataset 2}
}
\value{
mergedFile merged file from two datasets
}
\description{
merge sample scores from two datasets
}
\examples{
run_04_mergeTwoDatasets(merged_score1,merged_score2)
}
