% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_07_rankPathways.R
\name{run_07_rankPathways}
\alias{run_07_rankPathways}
\title{run_07_rankPathways}
\usage{
run_07_rankPathways(cca_res, p.info)
}
\arguments{
\item{cca_res}{CCA res}

\item{p.info}{pathway info}
}
\value{
rankP ranked pathways with info
}
\description{
rank pathways
}
\examples{
run_07_rankPathways()
}
