% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visObj.R
\name{visObj}
\alias{visObj}
\title{visObj}
\usage{
visObj(obj, dim_name, opfname)
}
\arguments{
\item{obj}{object}

\item{dim_name}{dim name to use 'umap' or 'cca'}

\item{opfname}{file name to save plot}
}
\description{
visualize obj using Seurat
}
\examples{
visObj(obj,dim_name,opfname)
}
